Allow repost w/ env var, fixes SD-2566
Also make build.sh work locally if you don't have all the vars set (end-
to-end tests that actually hit gerrit won't run) ... works in that it runs,
although it's below the coverage threshold so it fails at the end

Test Plan:
* specs? :fingerscrossed:

Reviewed-on: https://gerrit.instructure.com/115479
Tested-by: Jenkins
Product-Review: Zach Warrell <zwarrell@instructure.com>
QA-Review: Zach Warrell <zwarrell@instructure.com>
Tested-by: Zach Warrell <zwarrell@instructure.com>

+++ bin/run_tests.sh
@@ -35,17 +35,21 @@ run_command bundle exec rubocop
 export COVERAGE=1
 
 run_command bundle exec rspec
-run_command bin/gergich citest
-run_command bin/master_bouncer check
-DRY_RUN=1 run_command bin/master_bouncer check_all
 
-# ensure gergich works without .git directories
-rm -rf .git
-run_command bin/gergich status
+# these actually hit gerrit; only run them in CI land (you can do it
+# locally if you set all the docker-compose env vars)
+if [[ "$GERRIT_PATCHSET_REVISION" ]]; then
+  run_command bin/gergich citest
+  run_command bin/master_bouncer check
+  DRY_RUN=1 run_command bin/master_bouncer check_all
+  # ensure gergich works without .git directories
+  rm -rf .git
+  run_command bin/gergich status
+fi
 
 run_command bin/check_coverage
 
-if [[ $GEMNASIUM_TOKEN && $GEMNASIUM_ENABLED ]]; then
+if [[ "$GEMNASIUM_TOKEN" && "$GEMNASIUM_ENABLED" ]]; then
   # Push our dependency specification files to gemnasium for analysis
   run_command gemnasium dependency_files push -f=gergich.gemspec
 fi
